+++ kitty/launch.py
@@ -529,6 +529,9 @@ def load_watch_modules(watchers: Iterable[str]) -> Watchers | None:
         w = m.get('on_color_scheme_preference_change')
         if callable(w):
             ans.on_color_scheme_preference_change.append(w)
+        w = m.get('on_tab_bar_dirty')
+        if callable(w):
+            ans.on_tab_bar_dirty.append(w)
     return ans