+++ kitty/window.py
@@ -297,6 +297,7 @@ class Watchers:
     on_title_change: list[Watcher]
     on_cmd_startstop: list[Watcher]
     on_color_scheme_preference_change: list[Watcher]
+    on_tab_bar_dirty: list[Watcher]
 
     def __init__(self) -> None:
         self.on_resize = []
@@ -306,6 +307,7 @@ class Watchers:
         self.on_title_change = []
         self.on_cmd_startstop = []
         self.on_color_scheme_preference_change = []
+        self.on_tab_bar_dirty = []
 
     def add(self, others: 'Watchers') -> None:
         def merge(base: list[Watcher], other: list[Watcher]) -> None:
@@ -319,11 +321,13 @@ class Watchers:
         merge(self.on_title_change, others.on_title_change)
         merge(self.on_cmd_startstop, others.on_cmd_startstop)
         merge(self.on_color_scheme_preference_change, others.on_color_scheme_preference_change)
+        merge(self.on_tab_bar_dirty, others.on_tab_bar_dirty)
 
     def clear(self) -> None:
         del self.on_close[:], self.on_resize[:], self.on_focus_change[:]
         del self.on_set_user_var[:], self.on_title_change[:], self.on_cmd_startstop[:]
         del self.on_color_scheme_preference_change[:]
+        del self.on_tab_bar_dirty[:]
 
     def copy(self) -> 'Watchers':
         ans = Watchers()
@@ -334,12 +338,13 @@ class Watchers:
         ans.on_title_change = self.on_title_change[:]
         ans.on_cmd_startstop = self.on_cmd_startstop[:]
         ans.on_color_scheme_preference_change = self.on_color_scheme_preference_change[:]
+        ans.on_tab_bar_dirty = self.on_tab_bar_dirty[:]
         return ans
 
     @property
     def has_watchers(self) -> bool:
         return bool(self.on_close or self.on_resize or self.on_focus_change or self.on_color_scheme_preference_change
-                    or self.on_set_user_var or self.on_title_change or self.on_cmd_startstop)
+                    or self.on_set_user_var or self.on_title_change or self.on_cmd_startstop or self.on_tab_bar_dirty)
 
 
 def call_watchers(windowref: Callable[[], Optional['Window']], which: str, data: dict[str, Any]) -> None: